refactor: AuthService와 DogsService의 비동기 처리 성능 개선
- AuthService의 토큰 재발급 과정에서 Promise.all을 사용하여 병렬적으로 비동기 작업을 처리함.
- deleteUserData 메서드에서 사용자의 모든 개를 병렬적으로 삭제하도록 수정하고, 사용자 데이터 삭제와 S3 폴더 삭제를 병렬적으로 처리하도록 개선함.
- DogsService의 deleteDogFromUser 메서드에서 개와 관련된 데이터를 병렬적으로 삭제하도록 수정함.

+++ backend/src/auth/auth.service.ts
@@ -97,21 +97,15 @@ export class AuthService {
     async reissueTokens({ oauthId, provider }: RefreshTokenPayload): Promise<AuthData> {
         const { id: userId, oauthRefreshToken } = await this.usersService.findOne({ where: { oauthId } });
 
-        // TODO: Promise.all로 병렬처리 한 후 성능 비교하기
-        // const [
-        //     { access_token: newOauthAccessToken, refresh_token: newOauthRefreshToken },
-        //     newAccessToken,
-        //     newRefreshToken,
-        // ] = await Promise.all([
-        //     this[`${provider}Service`].requestTokenRefresh(oauthRefreshToken),
-        //     this.tokenService.signAccessToken(userId, provider),
-        //     this.tokenService.signRefreshToken(oauthId, provider),
-        // ]);
-        const { access_token: newOauthAccessToken, refresh_token: newOauthRefreshToken } =
-            await this[`${provider}Service`].requestTokenRefresh(oauthRefreshToken);
-
-        const newAccessToken = await this.tokenService.signAccessToken(userId, provider);
-        const newRefreshToken = await this.tokenService.signRefreshToken(oauthId, provider);
+        const [
+            { access_token: newOauthAccessToken, refresh_token: newOauthRefreshToken },
+            newAccessToken,
+            newRefreshToken,
+        ] = await Promise.all([
+            this[`${provider}Service`].requestTokenRefresh(oauthRefreshToken),
+            this.tokenService.signAccessToken(userId, provider),
+            this.tokenService.signRefreshToken(oauthId, provider),
+        ]);
 
         const attributes: Partial<Users> = { oauthAccessToken: newOauthAccessToken, refreshToken: newRefreshToken };
 
@@ -140,14 +134,9 @@ export class AuthService {
     private async deleteUserData(userId: number) {
         const dogIds = await this.usersService.getOwnDogsList(userId);
 
-        // TODO: for문 없애고 batch delete 하기
-        for (const dogId of dogIds) {
-            await this.dogsService.deleteDogFromUser(userId, dogId);
-        }
+        await Promise.all(dogIds.map((dogId) => this.dogsService.deleteDogFromUser(userId, dogId)));
 
-        // TODO: Promise.all로 병렬 처리
-        await this.usersService.delete({ id: userId });
-        await this.s3Service.deleteObjectFolder(userId);
+        await Promise.all([this.usersService.delete({ id: userId }), this.s3Service.deleteObjectFolder(userId)]);
     }
 
     async validateAccessToken(token: string): Promise<AccessTokenPayload> {
